refactor: changed installation path
